+++ CHANGELOG.md
@@ -2,6 +2,10 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ## [Unreleased]
 
+# Enhancements
+
+* Improve check for simpler programs
+
 ## v25.7.24
 
 ### Features